FriendFinderApp/requirements.txt
mysql-connector-python
tkinter